requirements.txt
"""This Code is built for pytorch GPU version. we highly recommend you to install pytorch for the GPU version!"""
pytorch==1.13.1
tqdm==4.66.2
pandas==2.0.3
torcheval==0.0.7
scikit-learn==1.3.2
